Add user authentication

+++ graphql/resolvers/booking.js
@@ -2,18 +2,21 @@ const Booking = require('../../models/Booking');
 
 const { transformBooking } = require('./merge');
 
-exports.bookings = async () => {
+exports.bookings = async (args, req) => {
+  if (!req.isAuth) throw new Error('Unauthorized');
   const bookingsResult = await Booking.find();
   return bookingsResult.map(bookingResult => transformBooking(bookingResult));
 };
-exports.bookEvent = async (args) => {
+exports.bookEvent = async (args, req) => {
+  if (!req.isAuth) throw new Error('Unauthorized');
   const bookingResult = await Booking.create({
     event: args.eventId,
-    user: '5c514026b71f4d2a5498a7d2',
+    user: req.userId,
   });
   return transformBooking(bookingResult);
 };
-exports.cancelBooking = async (args) => {
+exports.cancelBooking = async (args, req) => {
+  if (!req.isAuth) throw new Error('Unauthorized');
   const bookingResult = await Booking.findByIdAndDelete(args.bookingId);
   return transformBooking(bookingResult);
 };
